refactor: make openfeature.provider.registry a private module
